crypto/requirements.txt
pillow==11.2.1
pycryptodomex==3.22.0
